Refactor parse ansi str to iterator
